Refactor: 상벌점 태그 컴포넌트 분리
